+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/composable/ComposableMainScreen.kt
@@ -85,10 +85,12 @@ fun ComposableMainScreen(
                     }
                 }
             }
+
             ComposableInAppNotification(
                 modifier = Modifier.align(Alignment.BottomCenter),
                 notificationRenderer = notificationRenderer
             )
+            connectionApi.CheckAndShowUnsupportedDialog()
         }
     }
 }